+++ src/utils.ts
@@ -1,3 +1,5 @@
+import { DataFrame, FieldType, LinkModel } from '@grafana/data';
+import { findField } from '@volkovlabs/grafana-utils';
 import { Base64 } from 'js-base64';
 
 import { BASE64_MEDIA_HEADER_REGEX, IMAGE_TYPES_SYMBOLS } from './constants';
@@ -27,6 +29,25 @@ export const base64toBlob = (data: string, contentType: string, sliceSize = 512)
   }
 
   return new Blob(byteArrays, { type: contentType });
+};
+
+/**
+ * Get Data link for current value
+ * @param frames
+ * @param optionName
+ * @param currentIndex
+ */
+export const getDataLink = (frames: DataFrame[], optionName: string, currentIndex: number): LinkModel | null => {
+  const field = findField(
+    frames,
+    (field) => field.type === FieldType.string && (!optionName || field.name === optionName)
+  );
+
+  if (field && field?.getLinks) {
+    return field?.getLinks({ valueRowIndex: currentIndex })[0];
+  }
+
+  return null;
 };
 
 /**