+++ CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ## 5.2.0 (IN PROGRESS)
 
+### Breaking changes
+
+- Image links in panel options should be updated to use Data links.
+
 ### Features / Enhancements
 
 - Added poster for video (#120)
-- Update current index on data series decrease (#121)
+- Updated current index on data series decrease (#121)
+- Updated Image URL to data links (#122)
 
 ## 5.1.0 (2024-06-23)